+++ internal/cmd/toolhub/tools.go
@@ -78,7 +78,7 @@ func NewToolHubToolsListCmd() *cobra.Command {
 			out.Print(resultsToObjects(tools.Results, toolType))
 		},
 	}
-	out.AddFlags(cmd, string(output.TableType))
+	out.AddFlags(cmd, output.TableType)
 	cmd.Flags().StringVarP(&toolType, "type", "t", "*", "Type of tool: web app┃desktop app┃bot┃gadget┃user script┃command line tool┃coding framework┃other|\"\"")
 	return cmd
 }
@@ -107,7 +107,7 @@ func NewToolHubToolsSearchCmd() *cobra.Command {
 			out.Print(resultsToObjects(tools.Results, toolType))
 		},
 	}
-	out.AddFlags(cmd, "table")
+	out.AddFlags(cmd, output.TableType)
 	cmd.Flags().StringVarP(&toolType, "type", "t", "*", "Type of tool: web app┃desktop app┃bot┃gadget┃user script┃command line tool┃coding framework┃other|\"\"")
 	return cmd
 }